+++ packages/transport-grpc/CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+- 2022-03-16 -- [@bthaile](https://github.com/bthaile) Payer can now be an array.
 - 2022-02-11 -- Uses Buffer as provided by context injection
 
 ## 0.0.2 -- 2022-02-02